# utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hotel Management System",
    "description": "Hotel Management System API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3200/"
    }
  ],
  "schemes": ["http", "https"],
  "paths": {
    "/user": {
      "post": {
        "tags": ["auth"],
        "summary": "create a new User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/userInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/user/": {
      "get": {
        "tags": ["auth"],
        "summary": "get user by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header"
      }
    }
  },
  "definitions": {
    "userInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "required": "true"
        },
        "email": {
          "type": "string",
          "required": "true"
        },
        "phone": {
          "type": "number",
          "required": "true"
        },
        "role": {
          "type": "string",
          "required": "true"
        },
        "password": {
          "type": "string",
          "required": "true"
        }
      }
    },
...
  }
}